overleaf-remote-mcp/requirements.txt
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mcp==1.10.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
Werkzeug==3.1.3
